backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
pandas==2.2.3
openpyxl==3.1.5
pdfplumber==0.11.5
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.6.1
numpy==2.2.4
feedparser==6.0.11
holidays==0.71
psycopg2-binary==2.9.10
langchain==0.3.24
